refactor and move a bunch of stuff from connlib_shared to headless-client
