+++ rust/connlib/shared/Cargo.toml
@@ -41,9 +41,7 @@ hickory-proto = { workspace = true, optional = true }
 log = "0.4"
 
 [dev-dependencies]
-tempfile = "3.10.1"
 itertools = "0.12"
-mutants = "0.0.3" # Needed to mark functions as exempt from `cargo-mutants` testing
 tokio = { version = "1.36", features = ["macros", "rt"] }
 
 [target.'cfg(any(target_os = "macos", target_os = "ios"))'.dependencies]
@@ -51,10 +49,6 @@ swift-bridge = { workspace = true }
 
 [target.'cfg(target_os = "android")'.dependencies]
 tracing-android = "0.2"
-
-[target.'cfg(any(target_os = "linux", target_os = "windows"))'.dependencies]
-# Needed to safely backup `/etc/resolv.conf` and write the device ID on behalf of `gui-client`
-atomicwrites = "0.4.3"
 
 [target.'cfg(target_os = "linux")'.dependencies]
 netlink-packet-route = { version = "0.19", default-features = false }
