+++ rust/headless-client/src/lib.rs
@@ -34,6 +34,9 @@ use platform::default_token_path;
 /// Must be constructed inside a Tokio runtime context.
 use platform::Signals;
 
+/// Generate a persistent device ID, stores it to disk, and reads it back.
+pub(crate) mod device_id;
+pub mod dns_control;
 pub mod known_dirs;
 
 #[cfg(target_os = "linux")]
@@ -42,9 +45,9 @@ pub mod linux;
 pub use linux as platform;
 
 #[cfg(target_os = "windows")]
-pub mod windows;
+pub(crate) mod windows;
 #[cfg(target_os = "windows")]
-pub use windows as platform;
+pub(crate) use windows as platform;
 
 /// Only used on Linux
 pub const FIREZONE_GROUP: &str = "firezone-client";
@@ -57,7 +60,7 @@ pub const FIREZONE_GROUP: &str = "firezone-client";
 /// * `g` doesn't mean anything
 /// * `ed5437c88` is the Git commit hash
 /// * `-modified` is present if the working dir has any changes from that commit number
-pub const GIT_VERSION: &str = git_version::git_version!(
+pub(crate) const GIT_VERSION: &str = git_version::git_version!(
     args = ["--always", "--dirty=-modified", "--tags"],
     fallback = "unknown"
 );
@@ -83,7 +86,7 @@ struct Cli {
         env = "FIREZONE_API_URL",
         default_value = "wss://api.firezone.dev"
     )]
-    pub api_url: url::Url,
+    api_url: url::Url,
 
     /// Check the configuration and return 0 before connecting to the API
     ///
@@ -101,7 +104,7 @@ struct Cli {
     // AKA `device_id` in the Windows and Linux GUI clients
     // Generated automatically if not provided
     #[arg(short = 'i', long, env = "FIREZONE_ID")]
-    pub firezone_id: Option<String>,
+    firezone_id: Option<String>,
 
     /// Token generated by the portal to authorize websocket connection.
     // systemd recommends against passing secrets through env vars:
@@ -244,7 +247,7 @@ pub fn run_only_headless_client() -> Result<()> {
     // AKA "Device ID", not the Firezone slug
     let firezone_id = match cli.firezone_id {
         Some(id) => id,
-        None => connlib_shared::device_id::get().context("Could not get `firezone_id` from CLI, could not read it from disk, could not generate it and save it to disk")?.id,
+        None => device_id::get().context("Could not get `firezone_id` from CLI, could not read it from disk, could not generate it and save it to disk")?.id,
     };
 
     let (private_key, public_key) = keypair();
@@ -275,10 +278,11 @@ pub fn run_only_headless_client() -> Result<()> {
         rt.handle().clone(),
     );
     // TODO: this should be added dynamically
-    session.set_dns(platform::system_resolvers().unwrap_or_default());
+    session.set_dns(dns_control::system_resolvers().unwrap_or_default());
     platform::notify_service_controller()?;
 
     let result = rt.block_on(async {
+        let mut dns_controller = dns_control::DnsController::new();
         let mut tun_device = tun_device_manager::TunDeviceManager::new()?;
         let mut signals = Signals::new()?;
 
@@ -305,7 +309,7 @@ pub fn run_only_headless_client() -> Result<()> {
                     | InternalServerMsg::Ipc(IpcServerMsg::OnUpdateResources(_)) => {}
                     InternalServerMsg::OnSetInterfaceConfig { ipv4, ipv6, dns } => {
                         tun_device.set_ips(ipv4, ipv6).await?;
-                        tun_device.control_dns(dns).await?;
+                        dns_controller.set_dns(dns).await?;
                     }
                     InternalServerMsg::OnUpdateRoutes { ipv4, ipv6 } => {
                         tun_device.set_routes(ipv4, ipv6).await?
@@ -432,6 +436,7 @@ async fn handle_ipc_client(stream: platform::IpcStream) -> Result<()> {
     let (cb_tx, mut cb_rx) = mpsc::channel(10);
 
     let send_task = tokio::spawn(async move {
+        let mut dns_controller = dns_control::DnsController::new();
         let mut tun_device = tun_device_manager::TunDeviceManager::new()?;
 
         while let Some(msg) = cb_rx.recv().await {
@@ -439,7 +444,7 @@ async fn handle_ipc_client(stream: platform::IpcStream) -> Result<()> {
                 InternalServerMsg::Ipc(msg) => tx.send(serde_json::to_string(&msg)?.into()).await?,
                 InternalServerMsg::OnSetInterfaceConfig { ipv4, ipv6, dns } => {
                     tun_device.set_ips(ipv4, ipv6).await?;
-                    tun_device.control_dns(dns).await?;
+                    dns_controller.set_dns(dns).await?;
                     tx.send(serde_json::to_string(&IpcServerMsg::OnTunnelReady)?.into())
                         .await?;
                 }
@@ -461,8 +466,7 @@ async fn handle_ipc_client(stream: platform::IpcStream) -> Result<()> {
             IpcClientMsg::Connect { api_url, token } => {
                 let token = secrecy::SecretString::from(token);
                 assert!(connlib.is_none());
-                let device_id = connlib_shared::device_id::get()
-                    .context("Failed to read / create device ID")?;
+                let device_id = device_id::get().context("Failed to read / create device ID")?;
                 let (private_key, public_key) = keypair();
 
                 let login = LoginUrl::client(
@@ -473,7 +477,7 @@ async fn handle_ipc_client(stream: platform::IpcStream) -> Result<()> {
                     public_key.to_bytes(),
                 )?;
 
-                connlib = Some(connlib_client_shared::Session::connect(
+                let new_session = connlib_client_shared::Session::connect(
                     login,
                     Sockets::new(),
                     private_key,
@@ -481,7 +485,9 @@ async fn handle_ipc_client(stream: platform::IpcStream) -> Result<()> {
                     callback_handler.clone(),
                     Some(std::time::Duration::from_secs(60 * 60 * 24 * 30)),
                     tokio::runtime::Handle::try_current()?,
-                ));
+                );
+                new_session.set_dns(dns_control::system_resolvers().unwrap_or_default());
+                connlib = Some(new_session);
             }
             IpcClientMsg::Disconnect => {
                 if let Some(connlib) = connlib.take() {
