+++ rust/headless-client/src/linux.rs
@@ -3,15 +3,12 @@
 use super::{CliCommon, SignalKind, FIREZONE_GROUP, TOKEN_ENV_KEY};
 use anyhow::{bail, Context as _, Result};
 use connlib_client_shared::file_logger;
-use connlib_shared::linux::{etc_resolv_conf, get_dns_control_from_env, DnsControlMethod};
 use firezone_cli_utils::setup_global_subscriber;
 use futures::future::{select, Either};
 use std::{
-    net::IpAddr,
     os::unix::fs::PermissionsExt,
     path::{Path, PathBuf},
     pin::pin,
-    str::FromStr,
 };
 use tokio::{
     net::{UnixListener, UnixStream},
@@ -44,7 +41,7 @@ impl Signals {
     }
 }
 
-pub fn default_token_path() -> PathBuf {
+pub(crate) fn default_token_path() -> PathBuf {
     PathBuf::from("/etc")
         .join(connlib_shared::BUNDLE_ID)
         .join("token")
@@ -79,65 +76,6 @@ pub(crate) fn check_token_permissions(path: &Path) -> Result<()> {
         );
     }
     Ok(())
-}
-
-pub(crate) fn system_resolvers() -> Result<Vec<IpAddr>> {
-    match get_dns_control_from_env() {
-        None => get_system_default_resolvers_resolv_conf(),
-        Some(DnsControlMethod::EtcResolvConf) => get_system_default_resolvers_resolv_conf(),
-        Some(DnsControlMethod::NetworkManager) => get_system_default_resolvers_network_manager(),
-        Some(DnsControlMethod::Systemd) => get_system_default_resolvers_systemd_resolved(),
-    }
-}
-
-fn get_system_default_resolvers_resolv_conf() -> Result<Vec<IpAddr>> {
-    // Assume that `configure_resolv_conf` has run in `tun_linux.rs`
-
-    let s = std::fs::read_to_string(etc_resolv_conf::ETC_RESOLV_CONF_BACKUP)
-        .or_else(|_| std::fs::read_to_string(etc_resolv_conf::ETC_RESOLV_CONF))
-        .context("`resolv.conf` should be readable")?;
-    let parsed = resolv_conf::Config::parse(s).context("`resolv.conf` should be parsable")?;
-
-    // Drop the scoping info for IPv6 since connlib doesn't take it
-    let nameservers = parsed
-        .nameservers
-        .into_iter()
-        .map(|addr| addr.into())
-        .collect();
-    Ok(nameservers)
-}
-
-#[allow(clippy::unnecessary_wraps)]
-fn get_system_default_resolvers_network_manager() -> Result<Vec<IpAddr>> {
-    tracing::error!("get_system_default_resolvers_network_manager not implemented yet");
-    Ok(vec![])
-}
-
-/// Returns the DNS servers listed in `resolvectl dns`
-pub fn get_system_default_resolvers_systemd_resolved() -> Result<Vec<IpAddr>> {
-    // Unfortunately systemd-resolved does not have a machine-readable
-    // text output for this command: <https://github.com/systemd/systemd/issues/29755>
-    //
-    // The officially supported way is probably to use D-Bus.
-    let output = std::process::Command::new("resolvectl")
-        .arg("dns")
-        .output()
-        .context("Failed to run `resolvectl dns` and read output")?;
-    if !output.status.success() {
-        anyhow::bail!("`resolvectl dns` returned non-zero exit code");
-    }
-    let output = String::from_utf8(output.stdout).context("`resolvectl` output was not UTF-8")?;
-    Ok(parse_resolvectl_output(&output))
-}
-
-/// Parses the text output of `resolvectl dns`
-///
-/// Cannot fail. If the parsing code is wrong, the IP address vec will just be incomplete.
-fn parse_resolvectl_output(s: &str) -> Vec<IpAddr> {
-    s.lines()
-        .flat_map(|line| line.split(' '))
-        .filter_map(|word| IpAddr::from_str(word).ok())
-        .collect()
 }
 
 /// The path for our Unix Domain Socket
@@ -225,37 +163,4 @@ pub(crate) fn notify_service_controller() -> Result<()> {
 #[allow(clippy::unnecessary_wraps)]
 pub(crate) fn setup_before_connlib() -> Result<()> {
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use std::net::IpAddr;
-
-    #[test]
-    fn parse_resolvectl_output() {
-        let cases = [
-            // WSL
-            (
-                r"Global: 172.24.80.1
-Link 2 (eth0):
-Link 3 (docker0):
-Link 24 (br-fc0b71997a3c):
-Link 25 (br-0c129dafb204):
-Link 26 (br-e67e83b19dce):
-",
-                [IpAddr::from([172, 24, 80, 1])],
-            ),
-            // Ubuntu 20.04
-            (
-                r"Global:
-Link 2 (enp0s3): 192.168.1.1",
-                [IpAddr::from([192, 168, 1, 1])],
-            ),
-        ];
-
-        for (i, (input, expected)) in cases.iter().enumerate() {
-            let actual = super::parse_resolvectl_output(input);
-            assert_eq!(actual, expected, "Case {i} failed");
-        }
-    }
 }
